Revert "Revert users.id to INT"
This reverts commit 189023775179f26cb10b507b89c63f83ef6f5012.

+++ api/src/main/java/com/codeforcommunity/exceptions/UserDoesNotExistException.java
@@ -6,7 +6,7 @@ import io.vertx.ext.web.RoutingContext;
 public class UserDoesNotExistException extends HandledException {
   private String identifierMessage;
 
-  public UserDoesNotExistException(int userId) {
+  public UserDoesNotExistException(long userId) {
     this.identifierMessage = "id = " + userId;
   }
