+++ common/src/main/java/com/codeforcommunity/auth/JWTData.java
@@ -4,15 +4,15 @@ import com.codeforcommunity.enums.PrivilegeLevel;
 
 public class JWTData {
 
-  private final Integer userId;
+  private final Long userId;
   private final PrivilegeLevel privilegeLevel;
 
-  public JWTData(Integer userId, PrivilegeLevel privilegeLevel) {
+  public JWTData(Long userId, PrivilegeLevel privilegeLevel) {
     this.userId = userId;
     this.privilegeLevel = privilegeLevel;
   }
 
-  public Integer getUserId() {
+  public Long getUserId() {
     return this.userId;
   }
